Split Discord Service into two files.
